+++ identity.go
@@ -1,19 +1,12 @@
 package dbft
 
 import (
-	"encoding"
 	"fmt"
 )
 
 type (
 	// PublicKey is a generic public key interface used by dbft.
-	PublicKey interface {
-		encoding.BinaryMarshaler
-		encoding.BinaryUnmarshaler
-
-		// Verify verifies if sig is indeed msg's signature.
-		Verify(msg, sig []byte) error
-	}
+	PublicKey any
 
 	// PrivateKey is a generic private key interface used by dbft.
 	PrivateKey interface {